Check whether a string contains any Greek character
Currently, the `onlyGreek` method offers this functionality but it will return a string with empty spaces, if the supplied string contains any of them.

This is fine, as `onlyGreek` is intended to do so (based on it's current unit tests). However, adds an extra step for use cases that we just need to know if a string contains any greek characters or not.