Leet String to Normal String - Second attempt to address #23
Based on the feedback received on #36 have seen that there is current support for leet symbols as well as a function that transforms a normal string to a leet string.

This is a second attempt to address #23 by implementing a method that transforms a string with leet symbols to a string that will contain the equivalent letters.